removed testing files
