--- index.html
@@ -1,66 +1,0 @@
-<!DOCTYPE html>
-<html>
-    <head>
-        <meta charset="utf-8">
-        <title>Andy's LRS and xAPI Fun</title>
-        <script type="text/javascript" src="./dist/xapiwrapper.min.js"></script>
-        <script type="text/javascript" src="./src/xapi-util.js"></script>
-
-        <script type="text/javascript" src="./src/mymodule.js"></script>
-
-    </head>
-    <body>
-        <div class="one">
-            <p>
-                This is a paragraph.
-            </p>
-            <p id="two">
-                Wait for it ...
-            </p>
-            <p id="three">
-                Thank you. Please drive through.
-            </p>
-        </div>
-        <script type="text/javascript">
-
-            //configure
-            var conf = {};
-            conf['endpoint'] = "https://lrs.adlnet.gov/xapi/";
-            try {
-                conf['auth'] = 'Basic ' + toBase64('iitsecDemo:1234');
-            } catch (e) {
-                log('Exception in Config trying to encode auth: ' + e);
-            }
-            ADL.XAPIWrapper.changeConfig(conf);
-
-            //create statement
-            var stmt = new ADL.XAPIStatement(
-                'mailto:andrew.creighton.ctr@adlnet.gov',
-                'http://adlnet.gov/expapi/verbs/listened',
-                'http://classical.music.adlnet.gov');
-            stmt.generateId();
-            ADL.XAPIWrapper.sendStatement(stmt);
-
-            //prints statement just sent
-            var print1 = function (stmt)
-            {
-                document.getElementById('two').innerHTML = JSON.stringify(stmt);
-            }
-
-            //prints a group of statements
-            var printMany = function (stmtArr)
-            {
-                document.getElementById('three').innerHTML = JSON.stringify(stmtArr).replace(/},{/g, '}<br><br>{');
-            }
-
-            //retrieve statement with this id
-            ADL.XAPIWrapper.getStatements({'statementId': stmt.id}, null, function (r) {var stmt = JSON.parse(r.response); print1(stmt)});
-
-            // //retieves statements with this verb
-            ADL.XAPIWrapper.getStatements({'verb': 'http://adlnet.gov/expapi/verbs/listened'}, null, function (r) { var statements = JSON.parse(r.response).statements;               printMany(statements);});
-
-            //the following line is for testing along with line 9
-            alert(mymodule.test());
-        </script>
-    </body>
-</html>
