+++ src/xapi-util.js
@@ -5,18 +5,18 @@
         }
         return undefined;
     };
-    
+
     var getSubStatementDisplay = function (o) {
-        
+
     };
-    
+
     ADL.xapiutil = {};
-    
+
     ADL.xapiutil.getLang = function () {
-        return navigator.language || navigator.browserLanguage 
+        return navigator.language || navigator.browserLanguage
             || navigator.systemLanguage || navigator.userLanguage || "en-US";
     };
-    
+
     ADL.xapiutil.getMoreStatements = function (iterations, callback) {
         var stmts = [];
         ADL.XAPIWrapper.getStatements(null,null, function getMore(r) {
@@ -24,8 +24,8 @@
             var res = JSON.parse(r.response);
             if (! res.statements) return;
             stmts = stmts.concat(res.statements);
-            
-            if (iterations-- <= 0) { 
+
+            if (iterations-- <= 0) {
                 callback(stmts);
             }
             else {
@@ -35,11 +35,11 @@
             }
         });
     };
-    
+
     ADL.xapiutil.getActorId = function (a) {
         return a.mbox || a.openid || a.mbox_sha1sum || a.account;
     };
-    
+
     ADL.xapiutil.getActorIdString = function (a) {
         var id = a.mbox || a.openid || a.mbox_sha1sum;
         if (! id) {
@@ -49,45 +49,123 @@
         }
         return id;
     };
-    
+
     ADL.xapiutil.getActorDisplay = function (a) {
         return a.name || ADL.xapiutil.getActorIdString(a);
     };
-    
-    ADL.xapiutil.getVerbDisplay = function (v) { 
+
+    ADL.xapiutil.getVerbDisplay = function (v) {
         if (!v) return;
         if (v.display) return v.display[ADL.xapiutil.getLang()] || v.id;
         return v.id;
     };
-    
+
     ADL.xapiutil.getObjectType = function (o) {
         return o.objectType || ((o.id) ? "Activity" : "Agent");
     };
-    
+
     ADL.xapiutil.getObjectId = function (o) {
         if (o.id) return o.id;
         var type = ADL.xapiutil.getObjectType(o);
         if (type === "Agent" || type == "Group") return ADL.xapiutil.getActorId(o);
         return undefined;
     };
-    
-    ADL.xapiutil.getObjectIdString = function (o) {
-        if (o.id) return o.id;
-        var type = ADL.xapiutil.getObjectType(o);
-        if (type === "Agent" || type == "Group") return ADL.xapiutil.getActorIdString(o);
-        else if (type == "SubStatement") return ADL.xapiutil.getActorId(o.actor) + ":" + o.verb.id + ":" + ADL.xapiutil.getObjectId(o.object);
-        return undefined;
-    };
-    
-    ADL.xapiutil.getObjectDisplay = function (o) {
-        var disp = getObjDefName(o) || o.name || o.id;
-        if (! disp) {
-            var type = ADL.xapiutil.getObjectType(o);
-            if (type === "Agent" || type == "Group") disp = ADL.xapiutil.getActorDisplay(o);
-            else if (type == "SubStatement") disp = ADL.xapiutil.getActorId(o.actor) + ":" + o.verb.id + ":" + ADL.xapiutil.getObjectId(o.object);
+    (function (obj) {
+        var ADL = obj;
+        if (typeof window !== 'undefined') {
+            ADL = window.ADL = obj.ADL || {};
         }
-        
-        return disp;
-    };
-        
-})(window.ADL = window.ADL || {});
+
+        var getObjDefName = function (o) {
+            if (o.definition && o.definition.name) {
+                return o.definition.name[ADL.xapiutil.getLang()];
+            }
+            return undefined;
+        };
+
+        var getSubStatementDisplay = function (o) {
+
+        };
+
+        ADL.xapiutil = {};
+
+        ADL.xapiutil.getLang = function () {
+            return "en-US";
+            return navigator.language || navigator.browserLanguage
+                || navigator.systemLanguage || navigator.userLanguage || "en-US";
+        };
+
+        ADL.xapiutil.getMoreStatements = function (iterations, callback) {
+            var stmts = [];
+            ADL.XAPIWrapper.getStatements(null,null, function getMore(r) {
+                if (! (r && r.response) ) return;
+                var res = JSON.parse(r.response);
+                if (! res.statements) return;
+                stmts = stmts.concat(res.statements);
+
+                if (iterations-- <= 0) {
+                    callback(stmts);
+                }
+                else {
+                    if (res.more && res.more !== "") {
+                        ADL.XAPIWrapper.getStatements(null, res.more, getMore);
+                    }
+                }
+            });
+        };
+
+        ADL.xapiutil.getActorId = function (a) {
+            return a.mbox || a.openid || a.mbox_sha1sum || a.account;
+        };
+
+        ADL.xapiutil.getActorIdString = function (a) {
+            var id = a.mbox || a.openid || a.mbox_sha1sum;
+            if (! id) {
+                if (a.account) id = a.account.homePage + ":" + a.account.name;
+                else if (a.member) id = "Anon Group " + a.member;
+                else id = 'unknown';
+            }
+            return id;
+        };
+
+        ADL.xapiutil.getActorDisplay = function (a) {
+            return a.name || ADL.xapiutil.getActorIdString(a);
+        };
+
+        ADL.xapiutil.getVerbDisplay = function (v) {
+            if (!v) return;
+            if (v.display) return v.display[ADL.xapiutil.getLang()] || v.id;
+            return v.id;
+        };
+
+        ADL.xapiutil.getObjectType = function (o) {
+            return o.objectType || ((o.id) ? "Activity" : "Agent");
+        };
+
+        ADL.xapiutil.getObjectId = function (o) {
+            if (o.id) return o.id;
+            var type = ADL.xapiutil.getObjectType(o);
+            if (type === "Agent" || type == "Group") return ADL.xapiutil.getActorId(o);
+            return undefined;
+        };
+
+        ADL.xapiutil.getObjectIdString = function (o) {
+            if (o.id) return o.id;
+            var type = ADL.xapiutil.getObjectType(o);
+            if (type === "Agent" || type == "Group") return ADL.xapiutil.getActorIdString(o);
+            else if (type == "SubStatement") return ADL.xapiutil.getActorId(o.actor) + ":" + o.verb.id + ":" + ADL.xapiutil.getObjectId(o.object);
+            return undefined;
+        };
+
+        ADL.xapiutil.getObjectDisplay = function (o) {
+            var disp = getObjDefName(o) || o.name || o.id;
+            if (! disp) {
+                var type = ADL.xapiutil.getObjectType(o);
+                if (type === "Agent" || type == "Group") disp = ADL.xapiutil.getActorDisplay(o);
+                else if (type == "SubStatement") disp = ADL.xapiutil.getActorId(o.actor) + ":" + o.verb.id + ":" + ADL.xapiutil.getObjectId(o.object);
+            }
+
+            return disp;
+        };
+
+    })(this);
